requirements.txt
Flask>=2.2.0
pytest>=7.0.0
pytest-flask>=1.2.0
Flask-WTF>=1.0.0
WTForms>=3.0.0
Flask-Login>=0.6.0
Flask-SQLAlchemy>=3.0.0
bootstrap-flask>=2.3.3
psycopg2-binary
python-dotenv
email-validator
requests
Flask-SocketIO
redis
